database/requirements.txt
asgiref==3.5.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.11.0
django-import-export==2.8.0
django-rest-authtoken==2.1.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.9.0
pdfkit==1.0.0
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
